requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-registration-redux==2.12
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
random-username==1.0.2
sqlparse==0.4.4
typing-extensions==4.8.0
